Created archival service including endpoints to get for archival and retrieval or deletion and update the state, also added events and handlers to email requestors. Created file retention policy trigger endpoint, event and handler. Also updated graphql types with ID rather than Int and String.

+++ src/breedgraph/service_layer/infrastructure/file_management.py
@@ -2,28 +2,36 @@ import logging
 
 import aiofiles
 import asyncio
+import hashlib
 
 from pathlib import Path
 from fastapi import UploadFile, HTTPException
+from typing import Callable
 
 from breedgraph.service_layer.infrastructure.state_store import AbstractStateStore
-from breedgraph.config import FILE_STORAGE_PATH, MAX_CONCURRENT_UPLOADS
+
+from breedgraph.config import (
+    FILE_STORAGE_PATH,
+    MAX_CONCURRENT_UPLOADS
+)
 from breedgraph.custom_exceptions import IllegalOperationError
 from breedgraph.domain.model.submissions import SubmissionStatus
 
-from typing import Callable
-
 logger = logging.getLogger(__name__)
 
+
 class FileManagementService:
-    def __init__(self, state_store: AbstractStateStore):
+    def __init__(
+            self,
+            state_store: AbstractStateStore
+    ):
         self.file_storage_path = Path(FILE_STORAGE_PATH)
         self.file_storage_path.mkdir(parents=True, exist_ok=True)
         self.state_store = state_store
         self.upload_semaphore = asyncio.Semaphore(MAX_CONCURRENT_UPLOADS)
 
     async def save_file(
-            self, 
+            self,
             uuid: str,
             file: UploadFile,
             on_complete: Callable | None = None,
@@ -41,15 +49,24 @@ class FileManagementService:
                 file_size = file.size
                 logger.debug(f"Saving file {file.filename} with size {file_size} bytes as {uuid} ")
                 await self.state_store.set_status(uuid, SubmissionStatus.PROCESSING)
+
+                # Calculate hash while writing
+                hash_obj = hashlib.sha256()
                 async with aiofiles.open(file_path, 'wb') as buffer:
                     bytes_written = 0
                     while chunk := await file.read(8192):
                         await buffer.write(chunk)
+                        hash_obj.update(chunk)
                         bytes_written += len(chunk)
-                        progress = int((bytes_written/file_size) * 100)
+                        progress = int((bytes_written / file_size) * 100)
                         await self.state_store.set_file_progress(uuid, progress)
+
+                file_hash = hash_obj.hexdigest()
+                logger.debug(f"File {uuid} saved with hash {file_hash}")
+
                 if on_complete:
-                    await on_complete(uuid)
+                    await on_complete(file_size=file_size, file_hash=file_hash)
+
                 await self.state_store.set_status(uuid, SubmissionStatus.COMPLETED)
 
             except Exception as e:
@@ -59,18 +76,17 @@ class FileManagementService:
                 if on_failed:
                     await on_failed(uuid)
 
-    async def delete_file(
-            self,
-            uuid: str
-    ):
+    async def delete_file(self, uuid: str) -> None:
+        """Delete file locally """
         if not uuid:
             raise ValueError("UUID cannot be empty")
+
         file_path = Path(self.file_storage_path, uuid)
-        logger.debug(f"Deleting file {file_path}")
 
-        if file_path.is_dir():
-            # just another check to ensure pathlib isn't pointing at the file_storage_path still
-            # this should all be protected at the file system level with write access only within the file_storage_path
-            raise IllegalOperationError(f"Attempt to remove directory {file_path}, uuid supplied is empty string!")
-
-        file_path.unlink(missing_ok=True)
+        if file_path.exists():
+            logger.debug(f"Deleting file {file_path}")
+            if file_path.is_dir():
+                raise IllegalOperationError(
+                    f"Attempt to remove directory {file_path}, uuid supplied is empty string!"
+                )
+            file_path.unlink(missing_ok=True)
